auto_backend/requirements.txt
flask==2.2
flask-sqlalchemy==3.0.2
sqlalchemy==1.4.41
flask-cors==3.0.10
flask-migrate==3.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pandas==1.3.3